layers: Split Pipeline and Sampler state into separate files
This helps clean up some of the circular include dangers with
descriptor_sets.h, at the expense of a dinky sampler_state.h file.
