+++ BUILD.gn
@@ -96,7 +96,9 @@ core_validation_sources = [
   "layers/buffer_state.h",
   "layers/image_state.h",
   "layers/image_state.cpp",
+  "layers/pipeline_state.h",
   "layers/ray_tracing_state.h",
+  "layers/sampler_state.h",
   "layers/core_error_location.cpp",
   "layers/core_error_location.h",
   "layers/descriptor_sets.cpp",
